Rename files for WS
